chore: move neuron into modules
