+++ src/chop/nn/snn/modules/neuron/__init__.py
@@ -1,0 +1,2 @@
+from .neuron import BaseNode
+from .ifnode import IFNode
