+++ src/chop/nn/snn/modules/neuron/neuron.py
@@ -1,0 +1,217 @@
+# ***************************************************************************************
+# *    Title: neuron.py
+# *    Reference:  This file is adapted from spikingJelly
+# *    Availability: https://github.com/fangwei123456/spikingjelly/blob/master/spikingjelly/timing_based/neuron.py
+# *    Date: 07/11/2024
+# *    Code version: 0.0.0.014
+# ***************************************************************************************
+
+from abc import abstractmethod
+from typing import Callable, Optional
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import math
+import numpy as np
+import logging
+
+from ... import base
+from .. import surrogate
+
+from ...auto_cuda import neuron_kernel as ac_neuron_kernel
+from ...auto_cuda import ss_neuron_kernel as ss_ac_neuron_kernel
+
+try:
+    from ... import neuron_kernel, cuda_utils
+
+except BaseException as e:
+    logging.info(f"spikingjelly.activation_based.neuron: {e}")
+    neuron_kernel = None
+    cuda_utils = None
+
+
+class BaseNode(base.MemoryModule):
+    def __init__(
+        self,
+        v_threshold: float = 1.0,
+        v_reset: Optional[float] = 0.0,
+        surrogate_function: Callable = surrogate.Sigmoid(),
+        detach_reset: bool = False,
+        step_mode="s",
+        backend="torch",
+        store_v_seq: bool = False,
+    ):
+        """
+        * :ref:`API in English <BaseNode.__init__-en>`
+        .. _BaseNode.__init__-en:
+
+        :param v_threshold: threshold of this neurons layer
+        :type v_threshold: float
+
+        :param v_reset: reset voltage of this neurons layer. If not ``None``, the neuron's voltage will be set to ``v_reset``
+            after firing a spike. If ``None``, the neuron's voltage will subtract ``v_threshold`` after firing a spike
+        :type v_reset: Optional[float]
+
+        :param surrogate_function: the function for calculating surrogate gradients of the heaviside step function in backward
+        :type surrogate_function: Callable
+
+        :param detach_reset: whether detach the computation graph of reset in backward
+        :type detach_reset: bool
+
+        :param step_mode: the step mode, which can be `s` (single-step) or `m` (multi-step)
+        :type step_mode: str
+
+        :param backend: backend fot this neurons layer. Different ``step_mode`` may support for different backends. The user can
+        print ``self.supported_backends`` and check what backends are supported by the current ``step_mode``. If supported,
+        using ``'cupy'`` backend will have the fastest training speed
+        :type backend: str
+
+        :param store_v_seq: when using ``step_mode = 'm'`` and given input with ``shape = [T, N, *]``, this option controls
+            whether storing the voltage at each time-step to ``self.v_seq`` with ``shape = [T, N, *]``. If set to ``False``,
+            only the voltage at last time-step will be stored to ``self.v`` with ``shape = [N, *]``, which can reduce the
+            memory consumption
+        :type store_v_seq: bool
+
+        This class is the base class of differentiable spiking neurons.
+        """
+        assert isinstance(v_reset, float) or v_reset is None
+        assert isinstance(v_threshold, float)
+        assert isinstance(detach_reset, bool)
+        super().__init__()
+
+        if v_reset is None:
+            self.register_memory("v", 0.0)
+        else:
+            self.register_memory("v", v_reset)
+
+        self.v_threshold = v_threshold
+        self.v_reset = v_reset
+
+        self.detach_reset = detach_reset
+        self.surrogate_function = surrogate_function
+
+        self.step_mode = step_mode
+        self.backend = backend
+
+        self.store_v_seq = store_v_seq
+
+        # used in lava_exchange
+        self.lava_s_cale = 1 << 6
+
+        # used for cupy backend
+        self.forward_kernel = None
+        self.backward_kernel = None
+
+    @property
+    def store_v_seq(self):
+        return self._store_v_seq
+
+    @store_v_seq.setter
+    def store_v_seq(self, value: bool):
+        self._store_v_seq = value
+        if value:
+            if not hasattr(self, "v_seq"):
+                self.register_memory("v_seq", None)
+
+    @staticmethod
+    @torch.jit.script
+    def jit_hard_reset(v: torch.Tensor, spike: torch.Tensor, v_reset: float):
+        v = (1.0 - spike) * v + spike * v_reset
+        return v
+
+    @staticmethod
+    @torch.jit.script
+    def jit_soft_reset(v: torch.Tensor, spike: torch.Tensor, v_threshold: float):
+        v = v - spike * v_threshold
+        return v
+
+    @abstractmethod
+    def neuronal_charge(self, x: torch.Tensor):
+        """
+         * :ref:`API in English <BaseNode.neuronal_charge-en>`
+        .. _BaseNode.neuronal_charge-en:
+
+
+        Define the charge difference equation. The sub-class must implement this function.
+        """
+        raise NotImplementedError
+
+    def neuronal_fire(self):
+        """
+        * :ref:`API in English <BaseNode.neuronal_fire-en>`
+
+        .. _BaseNode.neuronal_fire-en:
+
+
+        Calculate out spikes of neurons by their current membrane potential and threshold voltage.
+        """
+
+        return self.surrogate_function(self.v - self.v_threshold)
+
+    def neuronal_reset(self, spike):
+        """
+        * :ref:`API in English <BaseNode.neuronal_reset-en>`
+
+        .. _BaseNode.neuronal_reset-en:
+
+
+        Reset the membrane potential according to neurons' output spikes.
+        """
+        if self.detach_reset:
+            spike_d = spike.detach()
+        else:
+            spike_d = spike
+
+        if self.v_reset is None:
+            # soft reset
+            self.v = self.jit_soft_reset(self.v, spike_d, self.v_threshold)
+
+        else:
+            # hard reset
+            self.v = self.jit_hard_reset(self.v, spike_d, self.v_reset)
+
+    def extra_repr(self):
+        return f"v_threshold={self.v_threshold}, v_reset={self.v_reset}, detach_reset={self.detach_reset}, step_mode={self.step_mode}, backend={self.backend}"
+
+    def single_step_forward(self, x: torch.Tensor):
+        """
+
+        * :ref:`API in English <BaseNode.single_step_forward-en>`
+
+        .. _BaseNode.single_step_forward-en:
+
+        :param x: increment of voltage inputted to neurons
+        :type x: torch.Tensor
+
+        :return: out spikes of neurons
+        :rtype: torch.Tensor
+
+        Forward by the order of `neuronal_charge`, `neuronal_fire`, and `neuronal_reset`.
+
+        """
+        self.v_float_to_tensor(x)
+        self.neuronal_charge(x)
+        spike = self.neuronal_fire()
+        self.neuronal_reset(spike)
+        return spike
+
+    def multi_step_forward(self, x_seq: torch.Tensor):
+        T = x_seq.shape[0]
+        y_seq = []
+        if self.store_v_seq:
+            v_seq = []
+        for t in range(T):
+            y = self.single_step_forward(x_seq[t])
+            y_seq.append(y)
+            if self.store_v_seq:
+                v_seq.append(self.v)
+
+        if self.store_v_seq:
+            self.v_seq = torch.stack(v_seq)
+
+        return torch.stack(y_seq)
+
+    def v_float_to_tensor(self, x: torch.Tensor):
+        if isinstance(self.v, float):
+            v_init = self.v
+            self.v = torch.full_like(x.data, v_init)
